cpp/2019/intcode: change `n_params` return type: `size_t` -> `int`
The `size_t` was polluting stuff with warnings here and there. I am
sure this change can break this code on some platforms, but I don't
really care right now.

+++ cpp/year2019/intcode/intcode.cc
@@ -155,8 +155,8 @@ namespace intcode {
     return m;
   }
 
-  size_t n_params(int opcode) {
-    size_t n;
+  int n_params(int opcode) {
+    int n;
     switch (opcode) {
     // addition, multiplication, less than, equals
     case 1:
